move files to src & tests folders, add util functions replaceParameterStringsInJSONValueWithKeepTags, translateRecursive
